measurement-sequence-system/requirements.txt
# Messsequenz-System - Python Dependencies
# VOLLSTÄNDIG AKTUALISIERT

# GUI
# tkinter ist in Python enthalten

# Wissenschaftliche Berechnungen
numpy>=1.21.0
scipy>=1.7.0

# Datenvisualisierung
matplotlib>=3.4.0

# Bildverarbeitung
Pillow>=9.0.0

# Datenverarbeitung
pandas>=1.3.0

# GUI-Automation und Steuerung
pyautogui>=0.9.53
pynput>=1.7.6

# Windows-spezifisch (nur unter Windows installieren)
# pywin32>=305  # Für Windows-Fenster-Steuerung

# Serielle Kommunikation (für echte Hardware)
pyserial>=3.5

# USB-Geräte
pyusb>=1.2.1

# Netzwerk-Kommunikation
requests>=2.26.0

# Zusätzliche Tools
python-dateutil>=2.8.0

# Optional für erweiterte Funktionen
opencv-python>=4.5.0  # Für erweiterte Bildverarbeitung
#keyboard>=0.13.5  # Alternative für Tastatur-Steuerung
